[BE#212] mates 서비스에 대한 테스트 코드 작성 (#306)
* test: mate service 테스트 코드 작성

* fix: 친구 통계 조회 시 배열 길이 6개인 오류 수정

* test: getMateAndMyStats() 테스트 케이스 추가

+++ BE/src/mates/mates.service.ts
@@ -85,7 +85,7 @@ export class MatesService {
     const result = await this.matesRepository.find({
       where: { follower_id: { id: user_id } },
     });
-    const userIds = result.map(({ following_id: { id } }) => id);
+    const userIds = result.map((following) => following.following_id);
     return Promise.all(
       userIds.map(async (id) => {
         const started_at = await this.redisService.get(`${id}`);
@@ -131,12 +131,17 @@ export class MatesService {
     const following = await this.userRepository.findOne({
       where: { id: following_id },
     });
+
+    if (!following) {
+      throw new NotFoundException('해당 유저는 존재하지 않습니다.');
+    }
+
     const result = await this.matesRepository.delete({
       follower_id: user,
       following_id: following,
     });
 
-    if (result.affected === 0) {
+    if (!result) {
       throw new NotFoundException('해당 친구 관계는 존재하지 않습니다.');
     }
   }
